lint

+++ plugin/evm/export_tx_test.go
@@ -154,7 +154,7 @@ func TestExportTxEVMStateTransfer(t *testing.T) {
 					Nonce:   0,
 				},
 			},
-			avaxBalance: uint256.NewInt(uint64(avaxAmount/2) * x2cRateUint64),
+			avaxBalance: uint256.NewInt(avaxAmount / 2 * x2cRateUint64),
 			balances: map[ids.ID]*big.Int{
 				customAssetID: big.NewInt(int64(customAmount)),
 			},
